services/transcribe/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
openai-whisper==20231117
torch==2.1.2
pydantic==2.5.3
httpx==0.26.0
aiofiles==23.2.1
ffmpeg-python==0.2.0
